Displayed First 5 Rows of Flight Data
Used the 'head()' function to display the first 5 rows of the flight data stored in the 'data' DataFrame. This preview allows initial inspection of the dataset's structure, providing insights into the column names, data format, and initial observations, aiding in the subsequent data exploration and analysis for the Flight Ticket Price Prediction ML project.