Verified and Dropped Records with Incorrect Duration
Implemented code to calculate the difference between 'Arrival_Time' and 'Dep_Time' in minutes and compare it with the 'Duration' column in the 'data' DataFrame. Mismatched records, where the calculated time difference didn't match the specified duration, were identified and removed from the DataFrame. The 'Time_Difference' column, which was used for comparison, was dropped from the DataFrame since it is no longer needed.