Removed Additional_Info Column with 'No Info' Majority
Based on analysis, it was observed that the majority of flights in the dataset have 'No Info' as the value in the 'Additional_Info' column. Considering this, we decided to streamline the data and simplify the feature set. The 'Additional_Info' column was dropped from the 'data' DataFrame to enhance data clarity and analysis efficiency.