Visualized Preferred Airline, Source, Destination, and Total Stops
Performed data visualization to identify the most preferred attributes in the 'Airline', 'Source', 'Destination', and 'Total_Stops' columns. The analysis showed that 'Jet Airways' is the most preferred airline, with 'Delhi' and 'Chennai' being the most frequent flight sources. For destinations, 'Kolkata' and 'Cochin' emerged as the top choices. Moreover, the majority of flights have either 1 or 0 total stops, indicating direct or non-stop flights are commonly preferred.