Added Random Forest Regression Model
Implemented a Random Forest Regression model using the 'RandomForestRegressor' from scikit-learn. The model was trained with 500 estimators and a minimum of 3 samples required to split an internal node ('min_samples_split=3'). This model demonstrates promising performance with an RMSE of 0.4108, an R-squared value of 0.8339, and an accuracy of 97.54% on the test dataset. This regression model can be used for further analysis and predictions in the project.